# backend/docs/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Housing Society Management API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/visitors": {
      "post": {
        "tags": [
          "visitors"
        ],
        "summary": "Pre Approve",
        "operationId": "pre_approve_api_v1_visitors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitorPreApproveIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/visitors/mine": {
      "get": {
        "tags": [
          "visitors"
        ],
        "summary": "My Visitors",
        "operationId": "my_visitors_api_v1_visitors_mine_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_VisitorOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/visitors/{visitor_id}/status": {
      "patch": {
        "tags": [
          "visitors"
        ],
        "summary": "Resident Update Status",
        "operationId": "resident_update_status_api_v1_visitors__visitor_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "visitor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Visitor Id"
            }
          },
          {
            "name": "new_status",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/VisitorStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/visitors/guard-view": {
      "get": {
        "tags": [
          "visitors"
        ],
        "summary": "Guard View",
        "description": "Restricted projection only \u2014 Section 21 (no financial/resident data).",
        "operationId": "guard_view_api_v1_visitors_guard_view_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GuardVisitorOut"
                  },
                  "type": "array",
                  "title": "Response Guard View Api V1 Visitors Guard View Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/visitors/{visitor_id}/entry": {
      "post": {
        "tags": [
          "visitors"
        ],
        "summary": "Guard Mark Entry",
        "operationId": "guard_mark_entry_api_v1_visitors__visitor_id__entry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "visitor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Visitor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardVisitorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/visitors/{visitor_id}/exit": {
      "post": {
        "tags": [
          "visitors"
        ],
        "summary": "Guard Mark Exit",
        "operationId": "guard_mark_exit_api_v1_visitors__visitor_id__exit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "visitor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Visitor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardVisitorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GuardVisitorOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "property_house_number": {
            "type": "string",
            "title": "Property House Number"
          },
          "visitor_name": {
            "type": "string",
            "title": "Visitor Name"
          },
          "visitor_mobile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visitor Mobile"
          },
          "visit_date": {
            "type": "string",
            "format": "date",
            "title": "Visit Date"
          },
          "status": {
            "$ref": "#/components/schemas/VisitorStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "property_house_number",
          "visitor_name",
          "visitor_mobile",
          "visit_date",
          "status"
        ],
        "title": "GuardVisitorOut",
        "description": "Section 21 data-boundary: Guard sees only what's needed for\nvisitor/security purposes \u2014 no financial/private resident data."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_VisitorOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VisitorOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "Page[VisitorOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisitorOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "society_id": {
            "type": "string",
            "format": "uuid",
            "title": "Society Id"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "resident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Resident Id"
          },
          "visitor_name": {
            "type": "string",
            "title": "Visitor Name"
          },
          "visitor_mobile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visitor Mobile"
          },
          "visit_date": {
            "type": "string",
            "format": "date",
            "title": "Visit Date"
          },
          "status": {
            "$ref": "#/components/schemas/VisitorStatus"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "society_id",
          "property_id",
          "resident_id",
          "visitor_name",
          "visitor_mobile",
          "visit_date",
          "status",
          "purpose",
          "created_at"
        ],
        "title": "VisitorOut"
      },
      "VisitorPreApproveIn": {
        "properties": {
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "visitor_name": {
            "type": "string",
            "title": "Visitor Name"
          },
          "visitor_mobile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visitor Mobile"
          },
          "visit_date": {
            "type": "string",
            "format": "date",
            "title": "Visit Date"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          }
        },
        "type": "object",
        "required": [
          "property_id",
          "visitor_name",
          "visit_date"
        ],
        "title": "VisitorPreApproveIn"
      },
      "VisitorStatus": {
        "type": "string",
        "enum": [
          "PRE_APPROVED",
          "EXPECTED",
          "ENTERED",
          "EXITED",
          "CANCELLED"
        ],
        "title": "VisitorStatus"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}